requirements.txt
fastapi
uvicorn[standard]
pandas
db-sqlite3
fastapi-cache2[redis]